+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiConstants.h
@@ -1,27 +1,24 @@
-//
-// Copyright (C) 2018 Google, Inc.
-//
-// SampleAdapterMediatedNativeAppInstallAd.h
-// Sample Ad Network Adapter
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
 
-#import "SampleAdapterMediatedNativeAd.h"
+// Adapter version string
+static NSString *const kGADMAdapterInMobiVersion = @"7.2.7.0";
 
-/// A subclass of the SampleAdapterMediatedNativeAd that adopts the GADMediatedNativeAppInstallAd
-/// protocol.
-@interface SampleAdapterMediatedNativeAppInstallAd
-    : SampleAdapterMediatedNativeAd<GADMediatedNativeAppInstallAd>
+// Tapjoy SDK key
+static NSString *const kGADMAdapterInMobiAccountID = @"accountid";
 
-@end
+// Tapjoy placemnt name
+static NSString *const kGADMAdapterInMobiPlacementID = @"placementid";
+
+static NSString *const kGADMAdapterInMobiErrorDomain = @"com.google.mediation.inmobi";