+++ adapters/Maio/MaioAdapter/GADMediationAdapterMaio.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google LLC.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,18 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import Foundation;
+#import <Foundation/Foundation.h>
+@import GoogleMobileAds;
 
-#import "GADMChartboostExtras.h"
-
-/// The purpose of the GADMAdapterChartboostDataProvider protocol is to allow the singleton to
-/// interact with the adapter.
-@protocol GADMAdapterChartboostDataProvider<NSObject>
-
-/// Returns the Chartboost extras object.
-- (GADMChartboostExtras *)extras;
-
-/// Returns the Chartboost ad location.
-- (NSString *)getAdLocation;
+@interface GADMediationAdapterMaio : NSObject <GADMediationAdapter>
 
 @end